Move class PageCollection into Foundation\Kernel namespace
