+++ packages/framework/src/Foundation/Concerns/BaseFoundationCollection.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Collection;
  * Base class for the kernel auto-discovery collections.
  *
  * @see \Hyde\Foundation\Kernel\FileCollection
- * @see \Hyde\Foundation\PageCollection
+ * @see \Hyde\Foundation\Kernel\PageCollection
  * @see \Hyde\Foundation\RouteCollection
  * @see \Hyde\Framework\Testing\Unit\BaseFoundationCollectionTest
  */
