+++ packages/framework/src/Foundation/HydeKernel.php
@@ -8,6 +8,7 @@ use Hyde\Facades\Features;
 use Hyde\Foundation\Kernel\FileCollection;
 use Hyde\Foundation\Kernel\Filesystem;
 use Hyde\Foundation\Kernel\Hyperlinks;
+use Hyde\Foundation\Kernel\PageCollection;
 use Hyde\Support\Concerns\Serializable;
 use Hyde\Support\Contracts\SerializableContract;
 use Illuminate\Support\Traits\Macroable;
